refactor: 토큰 타입 및 스키마 enum 패키지 변경
