+++ space-d/src/main/java/com/dnd/spaced/core/auth/application/TokenService.java
@@ -6,7 +6,7 @@ import com.dnd.spaced.core.auth.application.exception.ExpiredTokenException;
 import com.dnd.spaced.core.auth.application.exception.RotationRefreshTokenMismatchException;
 import com.dnd.spaced.core.auth.domain.PrivateClaims;
 import com.dnd.spaced.core.auth.domain.TokenDecoder;
-import com.dnd.spaced.core.auth.domain.TokenType;
+import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import com.dnd.spaced.core.auth.domain.repository.RefreshTokenRotationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
