+++ space-d/src/main/java/com/dnd/spaced/core/auth/domain/TokenDecoder.java
@@ -1,5 +1,6 @@
 package com.dnd.spaced.core.auth.domain;
 
+import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import java.util.Optional;
 
 public interface TokenDecoder {
