+++ space-d/src/main/java/com/dnd/spaced/core/auth/domain/TokenEncoder.java
@@ -1,5 +1,6 @@
 package com.dnd.spaced.core.auth.domain;
 
+import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import java.time.LocalDateTime;
 
 public interface TokenEncoder {
