--- space-d/src/main/java/com/dnd/spaced/core/auth/domain/TokenScheme.java
@@ -1,5 +1,0 @@
-package com.dnd.spaced.core.auth.domain;
-
-public enum TokenScheme {
-    BEARER
-}
