+++ space-d/src/main/java/com/dnd/spaced/core/auth/domain/enums/TokenScheme.java
@@ -1,0 +1,5 @@
+package com.dnd.spaced.core.auth.domain.enums;
+
+public enum TokenScheme {
+    BEARER
+}
