+++ space-d/src/main/java/com/dnd/spaced/core/auth/domain/enums/TokenType.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.auth.domain;
+package com.dnd.spaced.core.auth.domain.enums;
 
 public enum TokenType {
     ACCESS, REFRESH
