+++ space-d/src/main/java/com/dnd/spaced/core/auth/infrastructure/JwtEncoder.java
@@ -1,7 +1,7 @@
 package com.dnd.spaced.core.auth.infrastructure;
 
 import com.dnd.spaced.core.auth.domain.TokenEncoder;
-import com.dnd.spaced.core.auth.domain.TokenType;
+import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import com.dnd.spaced.global.config.properties.TokenProperties;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
