+++ space-d/src/test/java/com/dnd/spaced/core/auth/infrastructure/JwtEncoderTest.java
@@ -2,7 +2,7 @@ package com.dnd.spaced.core.auth.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.dnd.spaced.core.auth.domain.TokenType;
+import com.dnd.spaced.core.auth.domain.enums.TokenType;
 import com.dnd.spaced.global.config.properties.TokenProperties;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayNameGeneration;
